Прошу выйти из main файла и мигрировать в ког
